# Improve dirty grid rendering performance.