Software (HW), Multi-threading off (1080p):
- OpenRCT2 Title sequence (first park):
-- develop: 50~ FPS
-- PR: 80~ FPS
- bpb:
-- develop: 28~ FPS
-- PR: 45~ FPS

Software (HW), Multi-threading on (1080p):
- OpenRCT2 Title sequence (first park):
-- develop: 145~ FPS
-- PR: 155~ FPS
- bpb:
-- develop: 90~ FPS
-- PR: 95~ FPS

Clearly some gains with multi-threading off, this is also something OpenLoco apparently does.