@AaronVanGeffen 2023-04-26
I remember we changed this in OpenLoco to fix a bug, even. See https://github.com/OpenLoco/OpenLoco/issues/744 and https://github.com/OpenLoco/OpenLoco/pull/761