@ZehMatt 2023-04-26
> Is it okay if I test this out?

ofc