@Gymnasiast 2023-04-26
With optimisations (again BPB at 2560 × 1440):

Without multithreading:
Develop: 30 FPS
PR: 38 FPS

With multithreading:
Develop: 50 FPS
PR: 56 FPS