@janisozaur 2023-04-26
Looks like it's worthy of a changelog entry