@ZehMatt 2023-04-26
> With optimisations (again BPB at 2560 × 1440):
> 
> Without multithreading: Develop: 30 FPS PR: 38 FPS
> 
> With multithreading: Develop: 50 FPS PR: 56 FPS

Thanks, we are almost back to 40 FPS (vanilla) at 2k so that's not bad I guess. Would still feel a lot better at 60+ or even 120 when one owns a high refresh rate monitor :eyes: