@duncanspumpkin 2023-04-27
Awe man i remember working for ages to work out this issue in openloco but that was to fix z fighting. I tried it out in openrct2 but as it didn't fix z fighting i didn't proceed any further with porting it.